+++ IsekaiMod/Content/Features/IsekaiProtagonist/BeachEpisode/InnerPower.cs
@@ -4,14 +4,15 @@ using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.FactLogic;
+using UnityEngine;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.BeachEpisode
 {
     class InnerPower
     {
+        private static readonly Sprite Icon_BurningRenewal = Resources.GetBlueprint<BlueprintFeature>("7cf2a6bf35c422e4ea219fcc2eb564f5").m_Icon;
         public static void Add()
         {
-            var Icon_BurningRenewal = Resources.GetBlueprint<BlueprintFeature>("7cf2a6bf35c422e4ea219fcc2eb564f5").m_Icon;
             var InnerPower = Helpers.CreateBlueprint<BlueprintFeature>("InnerPower", bp => {
                 bp.SetName("Inner Power");
                 bp.SetDescription("You gain immunity to shaken, frightened, cowering, fear, death effects, {g|Encyclopedia:Ability_Scores}ability score{/g} {g|Encyclopedia:Damage}damage{/g}, energy drain, and negative levels.");
@@ -44,6 +45,8 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.BeachEpisode
                 bp.Ranks = 1;
                 bp.IsClassFeature = true;
             });
+
+            BeachEpisodeSelection.AddToSelection(InnerPower);
         }
     }
 }